Fix Py3 issues with the "echo" server in unit test
The "echo" server that uwsgi runs failed in Py3 but uwsgi ignored the error and just returned an empty result with status 200. Running the same server app with gunicorn flushed out the actual bug. Why is uwsgi so sucky? Let's not use it in the future.

Also fixing line endings in the nginx.conf.jinja template to unix.